# stats: Skip slowRejects if we have a definitive answer from fastRejects.