Commit Message: Checking stat rejection using the stats matcher is ALWAYS a two step process. We first do a fast check using fastRejects, and then call slowRejects with the output of fastRejects. The slow check needs to happen only if the answer from fastRejects is not definitive, i.e. if it is `FastResult::NoMatch` and not one of `FastResult::Matches` or `FastResult::Rejects`. 

We were previously doing the slow check on a few occasions even when we had a definitive answer from fastRejects (for eg. for `FastResult::Matches`). This PR fixes that.

Result of `stats_matcher_impl_speed_test` benchmark with and without changes:
```
Without changes:
-------------------------------------------------------
Benchmark             Time             CPU   Iterations
-------------------------------------------------------
BM_Inclusion     119343 ns       119334 ns         5884
BM_Exclusion     121073 ns       121073 ns         5767

With changes:
-------------------------------------------------------
Benchmark             Time             CPU   Iterations
-------------------------------------------------------
BM_Inclusion      29523 ns        29523 ns        23658
BM_Exclusion      28243 ns        28242 ns        24387
```
Additional Description:
Risk Level: Low
Testing: Added tests
Docs Changes: N/A
Release Notes: N/A
Platform Specific Features: N/A