@pradeepcrao 2023-05-18
> Also see the benchmark test for that case:
> 
> https://github.com/envoyproxy/envoy/blob/f4812764c43df5872fd773b25d399f7ff8fa4403/test/common/stats/thread_local_store_test.cc#L1116
> 
> Can we add similar benchmark tests for acceptance and include the before/after results in the PR description? Thank you!

I tried implementing this, but didn't see any difference with my changes, which is expected. My change essentially prevents the slow check if a stat matches in the fast check, but this has no impact on memory due to needing to memoize the rejection. 

I instead added a speed test and these are the results:
```
Before (without changes)
-------------------------------------------------------
 Benchmark             Time             CPU   Iterations
 -------------------------------------------------------
 BM_Inclusion     311879 ns       311876 ns         2241
 BM_Exclusion     306633 ns       306630 ns         2309

After (with changes)
 -------------------------------------------------------
 Benchmark             Time             CPU   Iterations
 -------------------------------------------------------
 BM_Inclusion     203780 ns       203774 ns         3451
 BM_Exclusion     197393 ns       197391 ns         3513

```