@pradeepcrao 2023-05-18
> I left some small benchmark suggestions. After you do those I expect the delta will be larger.
> 
> Can you put them benchmark results into the PR description?

Done. PTAL