@pradeepcrao 2023-05-19
> /retest

`//test/extensions/filters/http/file_system_buffer:file_system_buffer_integration_test` is failing with the following failure when run using `ci/run_envoy_docker.sh 'ci/do_ci.sh bazel.asan`. Doesn't seem related in any way to my changes.

```
[ RUN      ] Protocols/FileSystemBufferIntegrationTest.NotFoundBodyBuffer/IPv6_HttpDownstream_HttpUpstreamHttpParserNghttp2NoDeferredProcessing
unknown file: Failure
C++ exception with description "cannot bind '[::1]:0': Cannot assign requested address" thrown in the test body.
[  FAILED  ] Protocols/FileSystemBufferIntegrationTest.NotFoundBodyBuffer/IPv6_HttpDownstream_HttpUpstreamHttpParserNghttp2NoDeferredProcessing, where GetParam() = 24-byte object <01-00 00-00 00-00 00-00 00-00 00-00 00-00 00-00 00-00 00-00 00-00 00-00> (145 ms)
[ RUN      ] Protocols/FileSystemBufferIntegrationTest.NotFoundBodyBuffer/IPv6_HttpDownstream_Http2UpstreamHttpParserNghttp2NoDeferredProcessing
unknown file: Failure
C++ exception with description "cannot bind '[::1]:0': Cannot assign requested address" thrown in the test body.
[  FAILED  ] Protocols/FileSystemBufferIntegrationTest.NotFoundBodyBuffer/IPv6_HttpDownstream_Http2UpstreamHttpParserNghttp2NoDeferredProcessing, where GetParam() = 24-byte object <01-00 00-00 00-00 00-00 01-00 00-00 00-00 00-00 00-00 00-00 00-00 00-00> (41 ms)
[ RUN      ] Protocols/FileSystemBufferIntegrationTest.NotFoundBodyBuffer/IPv6_HttpDownstream_Http2UpstreamHttpParserOghttp2NoDeferredProcessing
unknown file: Failure
C++ exception with description "cannot bind '[::1]:0': Cannot assign requested address" thrown in the test body.
[  FAILED  ] Protocols/FileSystemBufferIntegrationTest.NotFoundBodyBuffer/IPv6_HttpDownstream_Http2UpstreamHttpParserOghttp2NoDeferredProcessing, where GetParam() = 24-byte object <01-00 00-00 00-00 00-00 01-00 00-00 00-00 00-00 01-00 00-00 00-00 00-00> (38 ms)
[ RUN      ] Protocols/FileSystemBufferIntegrationTest.NotFoundBodyBuffer/IPv6_HttpDownstream_Http2UpstreamBalsaParserNghttp2NoDeferredProcessing
unknown file: Failure
C++ exception with description "cannot bind '[::1]:0': Cannot assign requested address" thrown in the test body.
[  FAILED  ] Protocols/FileSystemBufferIntegrationTest.NotFoundBodyBuffer/IPv6_HttpDownstream_Http2UpstreamBalsaParserNghttp2NoDeferredProcessing, where GetParam() = 24-byte object <01-00 00-00 00-00 00-00 01-00 00-00 01-00 00-00 00-00 00-00 00-00 00-00> (39 ms)
[ RUN      ] Protocols/FileSystemBufferIntegrationTest.NotFoundBodyBuffer/IPv6_HttpDownstream_Http2UpstreamBalsaParserOghttp2NoDeferredProcessing
unknown file: Failure
C++ exception with description "cannot bind '[::1]:0': Cannot assign requested address" thrown in the test body.
[  FAILED  ] Protocols/FileSystemBufferIntegrationTest.NotFoundBodyBuffer/IPv6_HttpDownstream_Http2UpstreamBalsaParserOghttp2NoDeferredProcessing, where GetParam() = 24-byte object <01-00 00-00 00-00 00-00 01-00 00-00 01-00 00-00 01-00 00-00 00-00 00-00> (38 ms)
[ RUN      ] Protocols/FileSystemBufferIntegrationTest.NotFoundBodyBuffer/IPv6_Http2Downstream_HttpUpstreamHttpParserNghttp2NoDeferredProcessing
unknown file: Failure
C++ exception with description "cannot bind '[::1]:0': Cannot assign requested address" thrown in the test body.
[  FAILED  ] Protocols/FileSystemBufferIntegrationTest.NotFoundBodyBuffer/IPv6_Http2Downstream_HttpUpstreamHttpParserNghttp2NoDeferredProcessing, where GetParam() = 24-byte object <01-00 00-00 01-00 00-00 00-00 00-00 00-00 00-00 00-00 00-00 00-00 00-00> (37 ms)
[ RUN      ] Protocols/FileSystemBufferIntegrationTest.NotFoundBodyBuffer/IPv6_Http2Downstream_HttpUpstreamHttpParserOghttp2NoDeferredProcessing
unknown file: Failure
C++ exception with description "cannot bind '[::1]:0': Cannot assign requested address" thrown in the test body.
[  FAILED  ] Protocols/FileSystemBufferIntegrationTest.NotFoundBodyBuffer/IPv6_Http2Downstream_HttpUpstreamHttpParserOghttp2NoDeferredProcessing, where GetParam() = 24-byte object <01-00 00-00 01-00 00-00 00-00 00-00 00-00 00-00 01-00 00-00 00-00 00-00> (37 ms)
[ RUN      ] Protocols/FileSystemBufferIntegrationTest.NotFoundBodyBuffer/IPv6_Http2Downstream_HttpUpstreamBalsaParserNghttp2NoDeferredProcessing
unknown file: Failure
C++ exception with description "cannot bind '[::1]:0': Cannot assign requested address" thrown in the test body.
[  FAILED  ] Protocols/FileSystemBufferIntegrationTest.NotFoundBodyBuffer/IPv6_Http2Downstream_HttpUpstreamBalsaParserNghttp2NoDeferredProcessing, where GetParam() = 24-byte object <01-00 00-00 01-00 00-00 00-00 00-00 01-00 00-00 00-00 00-00 00-00 00-00> (38 ms)
[ RUN      ] Protocols/FileSystemBufferIntegrationTest.NotFoundBodyBuffer/IPv6_Http2Downstream_HttpUpstreamBalsaParserOghttp2NoDeferredProcessing
unknown file: Failure
C++ exception with description "cannot bind '[::1]:0': Cannot assign requested address" thrown in the test body.
[  FAILED  ] Protocols/FileSystemBufferIntegrationTest.NotFoundBodyBuffer/IPv6_Http2Downstream_HttpUpstreamBalsaParserOghttp2NoDeferredProcessing, where GetParam() = 24-byte object <01-00 00-00 01-00 00-00 00-00 00-00 01-00 00-00 01-00 00-00 00-00 00-00> (37 ms)
[ RUN      ] Protocols/FileSystemBufferIntegrationTest.NotFoundBodyBuffer/IPv6_Http2Downstream_Http2UpstreamHttpParserNghttp2NoDeferredProcessing
unknown file: Failure
C++ exception with description "cannot bind '[::1]:0': Cannot assign requested address" thrown in the test body.
[  FAILED  ] Protocols/FileSystemBufferIntegrationTest.NotFoundBodyBuffer/IPv6_Http2Downstream_Http2UpstreamHttpParserNghttp2NoDeferredProcessing, where GetParam() = 24-byte object <01-00 00-00 01-00 00-00 01-00 00-00 00-00 00-00 00-00 00-00 00-00 00-00> (38 ms)
[ RUN      ] Protocols/FileSystemBufferIntegrationTest.NotFoundBodyBuffer/IPv6_Http2Downstream_Http2UpstreamHttpParserOghttp2NoDeferredProcessing
unknown file: Failure
C++ exception with description "cannot bind '[::1]:0': Cannot assign requested address" thrown in the test body.
[  FAILED  ] Protocols/FileSystemBufferIntegrationTest.NotFoundBodyBuffer/IPv6_Http2Downstream_Http2UpstreamHttpParserOghttp2NoDeferredProcessing, where GetParam() = 24-byte object <01-00 00-00 01-00 00-00 01-00 00-00 00-00 00-00 01-00 00-00 00-00 00-00> (38 ms)
```